Move LibraryFolderContentsService to services package
Some additional rename refactoring and adaptations to remove StructuredApp dependency.
